# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barberia",
        "description": "Servicio de API tipo SOA para implementación en una barberia",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/reservacion/nueva": {
            "post": {
                "tags": [
                    "Reservaciones"
                ],
                "operationId": "bc82f4e968a91d145118a62944b26965",
                "requestBody": {
                    "description": "Datos de la reservación para crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdUsuario": {
                                        "description": "ID del usuario asociado a la reservación",
                                        "type": "integer",
                                        "example": 44
                                    },
                                    "IdServicio": {
                                        "description": "ID del servicio",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "FechaReserva": {
                                        "description": "Fecha de la reservación",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-11-21 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha creado la reservación correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al crear la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al crear la reservación."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/reservacion/idReservacion/editar": {
            "put": {
                "tags": [
                    "Reservaciones"
                ],
                "summary": "Actualizar un registro de reservación específico",
                "operationId": "80bda7907d995c2a5c6523045dd80230",
                "parameters": [
                    {
                        "name": "idReservacion",
                        "in": "path",
                        "description": "ID de la reservación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la reservación para actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdUsuario": {
                                        "description": "ID del usuario asociado a la reservación",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "FechaReserva": {
                                        "description": "Fecha de la reservación",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-11-21 15:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La reservación ha sido actualizada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al actualizar la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al actualizar la reservación."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/reservacion/idReservacion": {
            "get": {
                "tags": [
                    "Reservaciones"
                ],
                "summary": "Mostrar una reservación específica",
                "operationId": "77e4d1d5d09377c9a1b1eec9bb170395",
                "parameters": [
                    {
                        "name": "idReservacion",
                        "in": "path",
                        "description": "ID de la reservación a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "IdReservacion": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "IdUsuario": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "FechaReserva": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-21 15:30:00"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-21T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-21T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al mostrar la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al mostrar la reservación."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reservacion/idReservacion/eliminar": {
            "delete": {
                "tags": [
                    "Reservaciones"
                ],
                "summary": "Eliminar una reservación específica",
                "operationId": "e71bcfc718681cb05583b4a6c959bf00",
                "parameters": [
                    {
                        "name": "idReservacion",
                        "in": "path",
                        "description": "ID de la reservación a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La reservación ha sido eliminada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al eliminar la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al eliminar la reservación."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}